Refactor: extract cursor interaction code
