+++ crates/egui/src/widgets/text_edit/builder.rs
@@ -2,11 +2,21 @@ use std::sync::Arc;
 
 #[cfg(feature = "accesskit")]
 use accesskit::Role;
+
 use epaint::text::{cursor::*, Galley, LayoutJob};
 
-use crate::{output::OutputEvent, *};
+use crate::{
+    output::OutputEvent,
+    text_edit::cursor_interaction::{is_word_char, mouse_selection},
+    *,
+};
 
-use super::{CCursorRange, CursorRange, TextEditOutput, TextEditState};
+use super::{
+    cursor_interaction::{
+        ccursor_next_word, ccursor_previous_word, find_line_start, move_single_cursor,
+    },
+    CCursorRange, CursorRange, TextEditOutput, TextEditState,
+};
 
 /// A text region that the user can edit the contents of.
 ///
@@ -1372,294 +1382,7 @@ fn on_key_press(
     }
 }
 
-fn move_single_cursor(cursor: &mut Cursor, galley: &Galley, key: Key, modifiers: &Modifiers) {
-    if cfg!(target_os = "macos") && modifiers.ctrl && !modifiers.shift {
-        match key {
-            Key::A => *cursor = galley.cursor_begin_of_row(cursor),
-            Key::E => *cursor = galley.cursor_end_of_row(cursor),
-            Key::P => *cursor = galley.cursor_up_one_row(cursor),
-            Key::N => *cursor = galley.cursor_down_one_row(cursor),
-            Key::B => *cursor = galley.cursor_left_one_character(cursor),
-            Key::F => *cursor = galley.cursor_right_one_character(cursor),
-            _ => (),
-        }
-        return;
-    }
-    match key {
-        Key::ArrowLeft => {
-            if modifiers.alt || modifiers.ctrl {
-                // alt on mac, ctrl on windows
-                *cursor = galley.from_ccursor(ccursor_previous_word(galley.text(), cursor.ccursor));
-            } else if modifiers.mac_cmd {
-                *cursor = galley.cursor_begin_of_row(cursor);
-            } else {
-                *cursor = galley.cursor_left_one_character(cursor);
-            }
-        }
-        Key::ArrowRight => {
-            if modifiers.alt || modifiers.ctrl {
-                // alt on mac, ctrl on windows
-                *cursor = galley.from_ccursor(ccursor_next_word(galley.text(), cursor.ccursor));
-            } else if modifiers.mac_cmd {
-                *cursor = galley.cursor_end_of_row(cursor);
-            } else {
-                *cursor = galley.cursor_right_one_character(cursor);
-            }
-        }
-        Key::ArrowUp => {
-            if modifiers.command {
-                // mac and windows behavior
-                *cursor = Cursor::default();
-            } else {
-                *cursor = galley.cursor_up_one_row(cursor);
-            }
-        }
-        Key::ArrowDown => {
-            if modifiers.command {
-                // mac and windows behavior
-                *cursor = galley.end();
-            } else {
-                *cursor = galley.cursor_down_one_row(cursor);
-            }
-        }
-
-        Key::Home => {
-            if modifiers.ctrl {
-                // windows behavior
-                *cursor = Cursor::default();
-            } else {
-                *cursor = galley.cursor_begin_of_row(cursor);
-            }
-        }
-        Key::End => {
-            if modifiers.ctrl {
-                // windows behavior
-                *cursor = galley.end();
-            } else {
-                *cursor = galley.cursor_end_of_row(cursor);
-            }
-        }
-
-        _ => unreachable!(),
-    }
-}
-
 // ----------------------------------------------------------------------------
-
-/// Handle clicking and/or dragging text.
-fn mouse_selection(
-    ui: &mut Ui,
-    response: &Response,
-    cursor_at_pointer: Cursor,
-    galley: &Galley,
-    allow_drag_to_select: bool,
-    cursor_state: &mut text_edit::state::TextCursorState,
-) {
-    let text = galley.text();
-
-    if response.double_clicked() {
-        // Select word:
-        let ccursor_range = select_word_at(text, cursor_at_pointer.ccursor);
-        cursor_state.set_range(Some(CursorRange {
-            primary: galley.from_ccursor(ccursor_range.primary),
-            secondary: galley.from_ccursor(ccursor_range.secondary),
-        }));
-    } else if response.triple_clicked() {
-        // Select line:
-        let ccursor_range = select_line_at(text, cursor_at_pointer.ccursor);
-        cursor_state.set_range(Some(CursorRange {
-            primary: galley.from_ccursor(ccursor_range.primary),
-            secondary: galley.from_ccursor(ccursor_range.secondary),
-        }));
-    } else if allow_drag_to_select {
-        if response.hovered() && ui.input(|i| i.pointer.any_pressed()) {
-            ui.memory_mut(|mem| mem.request_focus(response.id));
-            if ui.input(|i| i.modifiers.shift) {
-                if let Some(mut cursor_range) = cursor_state.range(galley) {
-                    cursor_range.primary = cursor_at_pointer;
-                    cursor_state.set_range(Some(cursor_range));
-                } else {
-                    cursor_state.set_range(Some(CursorRange::one(cursor_at_pointer)));
-                }
-            } else {
-                cursor_state.set_range(Some(CursorRange::one(cursor_at_pointer)));
-            }
-        } else if ui.input(|i| i.pointer.any_down()) && response.is_pointer_button_down_on() {
-            // drag to select text:
-            if let Some(mut cursor_range) = cursor_state.range(galley) {
-                cursor_range.primary = cursor_at_pointer;
-                cursor_state.set_range(Some(cursor_range));
-            }
-        }
-    }
-}
-
-fn select_word_at(text: &str, ccursor: CCursor) -> CCursorRange {
-    if ccursor.index == 0 {
-        CCursorRange::two(ccursor, ccursor_next_word(text, ccursor))
-    } else {
-        let it = text.chars();
-        let mut it = it.skip(ccursor.index - 1);
-        if let Some(char_before_cursor) = it.next() {
-            if let Some(char_after_cursor) = it.next() {
-                if is_word_char(char_before_cursor) && is_word_char(char_after_cursor) {
-                    let min = ccursor_previous_word(text, ccursor + 1);
-                    let max = ccursor_next_word(text, min);
-                    CCursorRange::two(min, max)
-                } else if is_word_char(char_before_cursor) {
-                    let min = ccursor_previous_word(text, ccursor);
-                    let max = ccursor_next_word(text, min);
-                    CCursorRange::two(min, max)
-                } else if is_word_char(char_after_cursor) {
-                    let max = ccursor_next_word(text, ccursor);
-                    CCursorRange::two(ccursor, max)
-                } else {
-                    let min = ccursor_previous_word(text, ccursor);
-                    let max = ccursor_next_word(text, ccursor);
-                    CCursorRange::two(min, max)
-                }
-            } else {
-                let min = ccursor_previous_word(text, ccursor);
-                CCursorRange::two(min, ccursor)
-            }
-        } else {
-            let max = ccursor_next_word(text, ccursor);
-            CCursorRange::two(ccursor, max)
-        }
-    }
-}
-
-fn select_line_at(text: &str, ccursor: CCursor) -> CCursorRange {
-    if ccursor.index == 0 {
-        CCursorRange::two(ccursor, ccursor_next_line(text, ccursor))
-    } else {
-        let it = text.chars();
-        let mut it = it.skip(ccursor.index - 1);
-        if let Some(char_before_cursor) = it.next() {
-            if let Some(char_after_cursor) = it.next() {
-                if (!is_linebreak(char_before_cursor)) && (!is_linebreak(char_after_cursor)) {
-                    let min = ccursor_previous_line(text, ccursor + 1);
-                    let max = ccursor_next_line(text, min);
-                    CCursorRange::two(min, max)
-                } else if !is_linebreak(char_before_cursor) {
-                    let min = ccursor_previous_line(text, ccursor);
-                    let max = ccursor_next_line(text, min);
-                    CCursorRange::two(min, max)
-                } else if !is_linebreak(char_after_cursor) {
-                    let max = ccursor_next_line(text, ccursor);
-                    CCursorRange::two(ccursor, max)
-                } else {
-                    let min = ccursor_previous_line(text, ccursor);
-                    let max = ccursor_next_line(text, ccursor);
-                    CCursorRange::two(min, max)
-                }
-            } else {
-                let min = ccursor_previous_line(text, ccursor);
-                CCursorRange::two(min, ccursor)
-            }
-        } else {
-            let max = ccursor_next_line(text, ccursor);
-            CCursorRange::two(ccursor, max)
-        }
-    }
-}
-
-fn ccursor_next_word(text: &str, ccursor: CCursor) -> CCursor {
-    CCursor {
-        index: next_word_boundary_char_index(text.chars(), ccursor.index),
-        prefer_next_row: false,
-    }
-}
-
-fn ccursor_next_line(text: &str, ccursor: CCursor) -> CCursor {
-    CCursor {
-        index: next_line_boundary_char_index(text.chars(), ccursor.index),
-        prefer_next_row: false,
-    }
-}
-
-fn ccursor_previous_word(text: &str, ccursor: CCursor) -> CCursor {
-    let num_chars = text.chars().count();
-    CCursor {
-        index: num_chars
-            - next_word_boundary_char_index(text.chars().rev(), num_chars - ccursor.index),
-        prefer_next_row: true,
-    }
-}
-
-fn ccursor_previous_line(text: &str, ccursor: CCursor) -> CCursor {
-    let num_chars = text.chars().count();
-    CCursor {
-        index: num_chars
-            - next_line_boundary_char_index(text.chars().rev(), num_chars - ccursor.index),
-        prefer_next_row: true,
-    }
-}
-
-fn next_word_boundary_char_index(it: impl Iterator<Item = char>, mut index: usize) -> usize {
-    let mut it = it.skip(index);
-    if let Some(_first) = it.next() {
-        index += 1;
-
-        if let Some(second) = it.next() {
-            index += 1;
-            for next in it {
-                if is_word_char(next) != is_word_char(second) {
-                    break;
-                }
-                index += 1;
-            }
-        }
-    }
-    index
-}
-
-fn next_line_boundary_char_index(it: impl Iterator<Item = char>, mut index: usize) -> usize {
-    let mut it = it.skip(index);
-    if let Some(_first) = it.next() {
-        index += 1;
-
-        if let Some(second) = it.next() {
-            index += 1;
-            for next in it {
-                if is_linebreak(next) != is_linebreak(second) {
-                    break;
-                }
-                index += 1;
-            }
-        }
-    }
-    index
-}
-
-fn is_word_char(c: char) -> bool {
-    c.is_ascii_alphanumeric() || c == '_'
-}
-
-fn is_linebreak(c: char) -> bool {
-    c == '\r' || c == '\n'
-}
-
-/// Accepts and returns character offset (NOT byte offset!).
-fn find_line_start(text: &str, current_index: CCursor) -> CCursor {
-    // We know that new lines, '\n', are a single byte char, but we have to
-    // work with char offsets because before the new line there may be any
-    // number of multi byte chars.
-    // We need to know the char index to be able to correctly set the cursor
-    // later.
-    let chars_count = text.chars().count();
-
-    let position = text
-        .chars()
-        .rev()
-        .skip(chars_count - current_index.index)
-        .position(|x| x == '\n');
-
-    match position {
-        Some(pos) => CCursor::new(current_index.index - pos),
-        None => CCursor::new(0),
-    }
-}
 
 fn decrease_indentation(ccursor: &mut CCursor, text: &mut dyn TextBuffer) {
     let line_start = find_line_start(text.as_str(), *ccursor);
